moved caching logic from SplashScreenActivity.java to Cache.java
